Wed, Jun 28, 2023, 5:37 PM +08:00

+++ credit/credit.c
@@ -3,12 +3,12 @@
 
 int checksum1();
 int checksum2();
-int get_number();
+long get_number();
 int main(void)
 {
-    int number = get_number();
+    long number = get_number();
     int checksum1 = checksum1();
     int ckecksum2 = checksum2();
     int checksum = checksum1 + checksum2();
-    
+
 }
